CLOUDERA-BUILD Support for enabling webserver SSL in CSD
This patch was originally written by Todd [1]; Adar added secureUrl (so that
CM will generate https web UI links) and removed enabledOptionality (because
it was set to the default value).

It's not perfect due to the capabilities provided by the CSD subsystem;
see OPSAPS-38062 for details.

Adar tested this manually on a cluster, ensuring that metrics were still
collected and following the generated web UI links.

1. https://gerrit.cloudera.org/#/c/5806

Change-Id: I88cde67883f067b29055aa6df22d2c24c7dfe6fa