tablet: make time-based flushing configurable
We currently prioritize flushes for any in-memory data which has been
sitting longer than 2 minutes. This patch changes the time threshold to
be a new experimental flag rather than a hard-coded constant.

I found this useful to tweak when I wanted to run some experiments
regarding the on-disk size of data. With this flag, I was able to set
the threshold low to ensure that all of my inserted data was quickly
flushed to disk rather than sitting around in MRS where I couldn't
measure it.

Change-Id: Ic9a546b82e5b5db76473455df1eab2769c443c87