Add CM health aggregation to Kudu
In 5.13, CM added a type of health check known as health aggregation,
which can be configured in the CSD.
See http://github.mtv.cloudera.com/Starship/cm_ext/wiki/Service-Descriptor-Language-Reference#healthaggregation.

This adds support for this to the Kudu CSD.

The master percentages were chosen so that
- with 1 master, if the master is down the Kudu service is bad health
- with 3 masters, if a master is down the service is concerning health;
if >= 2 are down it's bad
- with 5 masters, if 1 or 2 masters are down the service is concerning
health; if >= 3 are down it's bad

Tablet server percentages are more arbitrary since "percentage of tablet
servers not in good health / running" isn't a good proxy for the health
of the service. For example, if a cluster has 50 tablet servers, and 2
servers go down simultaneously (> 95% up), then with very high
probability those two servers host 2/3 replicas for some tablet, so that
tablet's table will be in bad health and one might say the health of the
service is concerning because some tables are not writable. On the other
hand, if one server goes down, and 4 hours later another goes down too,
then having 2/50 down is fine, since the first server to go down has
(probably) had its data rereplicated already, so there's little or no
impact on users and clients because only 4% of cluster capacity is MIA.
I chose 75% as the threshold for concerning health and 50% for bad
health, because e.g. if > 25% of tablet servers are not healthy,
somebody should definitely be looking at what's wrong.