rpc: add support for passing an open file descriptor across a unix socket
This adds RpcController::SendFileDescriptor() which allows an RPC client
to pass an open file to the server. The server can retrieve the file
from the RpcContext.

The eventual goal here is that a client can create a memfd file (or an
unlinked file in /dev/shm if memfd isn't supported) and pass it to a
tserver in a Scan RPC. The server can then write the scan results into
the file instead of passing the data back via the normal RPC sidecar
mechanism. Assuming the client has mmapped the file, this should results
in the elimination of a bunch of copies and load on the reactor threads.

This includes a benchmark that compares various methods for doing this
shared memory transport. Results from running the benchmark on various
systems show that the shared memory can be substantially faster than the
existing sidecar approach:

88-core E5-2699 v4 @ 2.20GHz, el7:

Change-Id: I89baff83fa85f9109429fdc29d6eefc9f0f8b4a2